backend/requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==5.0.0
black==24.8.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.0
Flask==3.1.2
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mypy==1.11.2
mypy_extensions==1.1.0
numpy==2.4.2
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
protobuf==6.33.5
psycopg-binary==3.2.1
pycparser==2.23
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
pypdf==6.6.2
pytest==8.3.2
python-docx==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.16.2
ruff==0.6.8
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.30.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
